+++ lib/constants/strings.dart
@@ -1,3 +1,5 @@
+import 'package:vocabhub/constants/const.dart';
+
 /// app string constants  from the UI go here in this page
 /// e.g
 const String GITHUB_ASSET_PATH = 'assets/github.png';
@@ -26,3 +28,6 @@ const String signInFailure = 'failed to sign in User, please try again later';
 
 const String ABOUT_TEXT =
     "Vocabhub is a free and open-source project designed to aid in expanding vocabulary and earning new words. Developed as a personal side project by a single individual, this app aims to provide a helpful resource for users to enhance their linguistic skills. Personally, I have found the words available on this platform to be extremely beneficial, and I hope that you will find them useful as well.\nThe platform encourages contributions from users like yourself to improve its quality and benefit others. You can contribute by adding new words, refining existing ones, or reporting any bugs you encounter. Your suggestions and feedback are highly appreciated, and you can provide them through the report section of the app.\nI am always seeking contributors to enhance the word database on the platform. Your contributions will play a vital role in improving the overall quality and usefulness of Vocabhub. Vocabhub is licensed under the MIT License on GitHub, and you can access the source code by following the link provided.\nThank you for using Vocabhub, and I eagerly await your valuable feedback.";
+
+const String ratingDescription =
+    "If you've been finding value in your experience with ${Constants.APP_TITLE}, we kindly request a moment of your time to rate the app on the Play Store. We genuinely appreciate your support and feedback, and it will only take a minute. Thank you for choosing ${Constants.APP_TITLE}!";